Adjust margin-top of page and headers
The margin-top of the main page is set twice, but one of them (where it is zero) appears to be ignored.
Another one sets it to 35px which seems like too much. This change makes the one that "takes effect" smaller.

Another change is for headers. If a header is a first element, it also gives a large margin at the top. This one
is harder to fix as it needs a css which I don't think is exactly available: "If the first child of the main panel
is an "a" and the second element of this panel is an h1/h2/h3, then make the margin top of that h* to be zero, but leave
all the margin-top settings of all subsequent h* elements".
I have gone with a quick fix, which seems like it will work in all cases, though it may now seems slightly closer to the top
than is necessary.